Back to floats

+++ source/camera.cpp
@@ -1,6 +1,6 @@
 #include "camera.h"
 
-Ray Camera::createRay(double u, double v) const
+Ray Camera::createRay(float u, float v) const
 {
 	Vec3 direction(u, v, 1.0);
 	direction.normalise();
